support create table syntax, format code
