+++ .babelrc
@@ -1,0 +1,18 @@
+{
+  "presets": [
+    [
+      "env",
+      {
+        "targets": {
+          "node": "10.16.3"
+        },
+        "modules": "commonjs",
+        "useBuiltIns": true,
+        "debug": false
+      }
+    ]
+  ],
+  "plugins": [
+    "transform-object-rest-spread"
+  ]
+}
