+++ .gitignore
@@ -1,3 +1,5 @@
 node_modules/
 build/
 coverage/
+lib/
+index.d.ts
