+++ src/constrain.js
@@ -1,0 +1,8 @@
+
+function constraintDefinitionToSQL(constraintDefinition) {
+  return constraintDefinition
+}
+
+export {
+  constraintDefinitionToSQL,
+}
