+++ src/interval.js
@@ -1,0 +1,8 @@
+function intervalToSQL(expr) {
+  const [intervalNum, unit] = expr.value
+  return `INTERVAL ${intervalNum} ${unit}`
+}
+
+export {
+  intervalToSQL,
+}
